Allow "\n", "\t" and ‘ ’ at the end of integer [for `q = int(lines[0])` in algoprog.ru/.../111580+.py]

+++ 11l_hpp/Array.hpp
@@ -667,7 +667,7 @@ template <typename Type> Type min(const Array<Type> &arr)
 template <class T, size_t N> struct array_binder {
 	//const Array<T> &arr; // this does not work with `auto [a, b] = bind_array<2>(create_array({1, 2}));` because destructor of Array is called before `get()`
 	T a[N];
-	array_binder(const Array<T> &arr) { for (size_t i=0; i<N; i++) a[i] = arr[i]; }
+	array_binder(const Array<T> &arr) { for (size_t i=0; i<N; i++) a[i] = arr[(int)i]; }
 	template <size_t I> T get() { return a[I]; }
 };
